requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-extensions==2.1.9
djangorestframework==3.9.4
idna==2.8
mysqlclient==1.4.2.post1
psycopg2==2.8.3
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
